Move FeedItemsMapper and HTTPClient to its own file
