+++ EssentialFeed/FeedAPI/FeedItemsMapper.swift
@@ -1,0 +1,40 @@
+//
+//  FeedItemsMapper.swift
+//  EssentialFeed
+//
+//  Created by Max Kai on 2023/1/21.
+//
+
+import Foundation
+
+internal final class FeedItemsMapper {
+    private struct Root: Decodable {
+        let items: [Item]
+    }
+
+    private struct Item: Decodable {
+        let id: UUID
+        let description: String?
+        let location: String?
+        let image: URL
+        
+        var item: FeedItem {
+            return FeedItem(
+                id: id,
+                description: description,
+                location: location,
+                imageURL: image)
+        }
+    }
+    
+    private static var OK200: Int { return 200 }
+    
+    internal static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [FeedItem] {
+        guard response.statusCode == OK200 else {
+            throw RemoteFeedLoader.Error.invalidData
+        }
+        
+        let root = try JSONDecoder().decode(Root.self, from: data)
+        return root.items.map { $0.item }
+    }
+}
